requirements.txt
PySide6
Pillow
opencv-python